Add new Reminder feature to Bobi
This reminder feature give users an overview of their tasks that are of
a higher priority so that they can focus on those tasks first.

A reminder will pop up when user first launch the Bobi GUI. It will
remind users of any overdue tasks and tasks due this coming week as they
are usually of higher priority.

This feature is very necessary. If the user has a long list of tasks to
do, going through the list every single time will be a very big
hassle. This feature helps them stay on track as it reduces the chances
of them missing deadlines.

Yet to do:
- Javadoc comments for newly added code
- code quality check for newly added code
- consider adding a command where users can view this reminder whenever
  they want