Add an undo feature to undo most recent change
There might be some cases where users will accidentally delete or change the content of the list.

In such cases, an undo feature will be really helpful in helping users retrieve lost or changed information.

The undo feature can only allow users to undo the most recent change and cannot be used to undo more than one change.

Furthermore, the undo feature will only detect mark, unmark, delete, and the three add task commands as commands that change the content of the list. After all, the other features such as list and find does not change the contents of the task list.

Finally, upon exiting and restarting the chatbot, the undo will not work. It will only work when a change is made in the current session.