Add assert feature
Assert feature is used to document important assumptions that should hold at various points in the code and check for the accuracy of the code.

It is added it to the Parser class as the class parse and process the input and gives the output, so its accuracy is very important.

Let's,
* the asserts checks if the methods returns a null.