Rename variable to code standard, breaking backwards data compatibility
To comply with coding standards, we rename an internal boolean variable
used by the Task class to a "boolean-sounding prefix".

This was originally not done because the variable was depended upon by
Gson, which meant changing the name will break backwards compatibility.
However, in light of automated grading used, which meant it might not
read the documentation with regards to variable name rationales, it is
decided that future builds from now on shall not be compatible with
past alpha builds, and thus the variable is renamed to avoid flagging
the grader.