Add TimedTask feature
Before there was not any feature to make and keep track of task which
can be done in a certain period of time without certainty of the start
and the end of the task

Enabling to add, mark, unmark, and remove TimedTask which is done in the
same way as the other type of tasks.

This way is used to maintain the uniform of code with the other type of
tasks.