Use assertions
Private methods in some classes have invariants for their arguments.

Such invariants should be checked to ensure assumptions made about
such arguments are right. They are typically to be tested at the
start of methods.

Assertions are used to check for logically impossible situations,
but they will not be used at runtime.