Improve Code Quality
Critically examine the code and refactor to improve the code quality in the Parser class.
The method parse(TaskList tasks, String echo) in the Parser class was too long and had arrow head shape, due to the use of embedded if else statements.
The long method is being broken down into many shorter methods that each does a function, according to SLAP. for example markCase(TaskList tasks, String[] parts) marks the case. The embedded if-else statements are also cleaned to reduce the arrow head effect.