Fix bugs regarding filepaths and edit code to write to file after each command.
Currently, the filepath was hardcoded and was unable to load from any directory. Writing to file only happened with bye command.

Changed the filepath to be relative to where the program is stored. If file not found, a new directory will be created. Program now writes to file after each command.

This is to ensure user can run jar file no matter where he stores it. Writing to file after each command ensures user progress is saved.