Add a Sort Command to sort the tasks.
The Sort Command is done by entering "sort".
It sorts the task by the type of the task initially,
based on this order: ToDo, Deadline, Event.
Then, for Deadline, it will sort by the deadline date chronologically.
Then, for Event, it will sort by the start date chronologically.
After sorting, the GUI will show the sorted list of tasks.