Added a new feature to direct people to a README
Added a new command "help" to direct users to my github page, where the README.md file is located

The README file will contain all of the instructions and guidance required for users to know what command they can use when utilizing Bob.

The "help" command is chosen as it is something intuitive for users to type in when they are unfamiliar with the command required.