Add help command
Because there are many commands, new users may be confused when
they use the app for the first time.

Add a help command to tell the purpose of the app and list out
all of the current commands as well as their formats.

Doing so will help new users feel more convinient whenever they
forget the syntax of any command.