Parser class: abstract away the validation of commands into the parsing methods
Parser class initially had 2 different operations - 1) To decide the type of command to be returned 2) Check for the validity of the formatting of the command once the type of command to be returned was determined.

Abstracting the second operation(validation check) allows the parse method to be simpler and look neater.