Improve colour appearance in GUI
The chatbot so far does not distinguish between error responses by the
bot and regular responses. When there is an error response, a dark red
reply will be shown by the bot. Dark red is chosen to improve contrast
and thus readability of the chatbot.

The background colours are changed to improve distinction between the
bot and the human input. The bot has a default colour palette (same as
background) while the human input has a silver colour palette.