Implement priority tracking of tasks in chatbot.
This is to enable users to track the priority of
completing tasks.

Priority enum is used to ensure that the priority
levels of the task falls within specific range.
SetPriority command is also added to handle the
update of priority.

Another implementation is the ability to save the
priority level in the text file so that its status
can be retrieved in the future.

Further extension can be done to give functionality
to the different priority levels.