Add assertions to confirm that inputs for Todo, Deadline or Event tasks are not null
Previously, if else statements were used to determine whether the input was valid (i.e. not null). This caused more nested code and reduced the code quality, and prevented us from checking for input validity efficiently.

Add assert statements to return an assertion error when the user inputs no arguments for a Todo, Deadline or Event task.

This allows for inputs to be checked in one line, and reduces nested code.