My last commit (hopefully) :)
The code does not handle the case where the data files do not exist at
the start. When someone else takes my chatbot and runs it for the first
time, the required data files will not exist in their computer

Let's change this by checking if those files exist, and create them if
they do not. This is done whenever the application is opened. Thus
users will always have the required data files to save their tasks and
commands to memory