Create new classes for commands
Currently, the Parser class is extremely long as most of the command methods
are contained within the class.

This makes the code extremely long and probes further abstraction and
simplification of the code.

Let's make new classes to encapsulate the commands!

This way, our code has a higher level of abstraction and is more organised.