Add detection to ensure Event start is before end
With the previous implementation, Event end date/time can be set to be
before the start date/time. With this detection in place, when creating
or updating an Event, there is a check in place to ensure Event from
date/time is before Event to date/time. This check is done in the
constructor, when updating using the `update()` method, as well as when
cloning the Event.