Improve error handling when parsing Events
Previously, when the parser parses an event command, the parser will
disregard the logic of /from and /to markers within the event command.

This fixes the issue and ensures that exactly 1 /from and 1 /to marker
is present in the event command and a few JUnit test cases under
parseEvent have been added.

JUnit test cases:
- 2 /from: DukeException thrown
- 2 /to: DukeException thrown
- /from and /to flipped: accepted and event is flipped as needed