Update access modifiers for Task, Deadline, Event
Previously, there were no access modifiers for some variables in the mentioned classes. This may lead to improper usage of private class variables.

Update the access modifiers for certain variables to be protected or private, as appropriate. Prevents unintentional breaking of the abstraction barrier.

As a step towards prevent unintentional breaking of the abstraction barrier and improper usage of supposedly private variables, let's be more mindful and remember to assign appropriate access variables when creating new variables and methods.