Shorten loadTaskFromFile method
loadTaskFromFile method needs to be shortened
and simplified to ensure greater readability and
standardised levels of abstraction. This makes
it challenging for people to understand the code.

Let’s create a createTask method that helps
to abstract out details specific to the creation
of tasks and handle them in another method.

This improves readability.