Shorten loadTaskFromFile method
loadTaskFromFile method needs to be shortened
And simplified to ensure greater readability and
standardised levels of abstraction. This makes
It challenging for people to understand the code.

Let’s create a updateTaskList method that helps
to abstract out details specific to the updating
of the task list when we load the data.

This improves readability and simplifies the code.