Add asserts to files
Unify variations of toSet() methods

There are several methods that finds the index of a task in the
TaskList. Thus, we need to check whether the given index is
valid.

Asserting the validity of said index ensures the correctness of
the index before finding the task itself.

Let's add asserts in the delete, mark and unmark methods in both
the TaskList and the Parser.