Ui: Remove the Scanner variable and import DialogBox: Remove the flip method
To abide to coding standards, and with an especially large
number of files, all unused variables and imports must be
removed.

Removing such unused variables saves space and streamlines
the code as per the CS2103T coding standards, and indicates
improvement.

Let's remove the unused scanner methods in the Ui and the
flip methods in DialogBox, as there is no longer a need of
CLI scanners with the GUI and there is already a separate
class responsible for showing the opposing chatbot.