Improve code quality
There are processes to follow when user inputs do not
fit expected parameters. In most cases, the error is thrown through
IllegalArgumentException.

Using a checked exception forces the case to be explicitly handled.

Defining and using a custom PhiException allows for greater control
as to what happens after unexpected user inputs occur.