Provide more help to users
Users only have access to a help command, which may be daunting
when first experiencing a new interface and app.

Pre-load the app with sample data at the first run, and polish the
help command.

This allows users to have a guideline as to how the app functions,
and the streamlined help command provides in-app guidance to them
regarding the different commands they are able to use in the app.