Break down Command class into separate classes
The Command class is an all-purpose class that encompasses all valid commands. The main execution of the commands is handled by the executeCommand method in the Duke class and the corresponding methods in TaskList.

Having specific classes for each command will allow the unique execution processes of each command to be decentralised.

Let's make Command be an abstract parent class with the execute method, and make individual classes for each valid command.

This means that all Commands can be run by Duke using the execute method, without having to know the details of how each command is being executed.