Add assertions to code
This commit introduces assertions to the codebase by enabling it in
`gradle.build`.

There are several methods in the code that should hold at various points
to ensure the safety of our program. This includes certain
uninstantialised objects which could be null values.

By using assertions, we are able to put assertion points that ensure
our variables are not null values.