Update readability and fix bugs
Unused methods in a few classes.

Since updating to a JavaFX GUI, some methods do not need to be used.

Let's remove unused methods in ExitCommand.java and update missing
JavaDocs in newer classes like EditCommand.java and newer methods like
editByDate() in Deadline.java.

Edit command: date edits not working

Edit command for dates are not working.

Let's refactor code in EditCommand to read the correct cases and add a
default case that throws error if the field is not recognised.