Handle data storage exceptions
Storage.java: Throws error if ./data/dre.txt not found.

Make Storage.java create the file itself if it doesn't exist.

Let's add methods to check if the storage file exists, and create the
file if it doesn't.

This allows the application to exit normally and the jar file now
operates independently of any other folder.