GUI: Add functionality for list to read and be initialised from existing file.
The list starts off as empty, and does not read from any file.

The data has to be added for each time ChatterBot is run, for the GUI.

Let's get the list to read the ChatterBot.txt file and be initialised with the current values inside.

Using this will allow the user to use ChatterBot over a period of time with the same tasks inside, instead of having to re-add them every time.

Still some bugs to fix, like for the 'delete' and 'mark'/'unmark' functions.