Refactor Parser class parse method
The parse method is too long and does not adhere to SLAP.

Extracting the bodies of the case statements into individual
functions would fix this issue and make the code more readable.

Let's create a method to parse the regex Matcher into a Command
for every branch of the switch case.