Refactor Storage class loadTasks method
The loadTasks method is too long and does not adhere to SLAP.

Extracting the bodies of the case statements into individual
functions would fix this issue and make the code more readable.

Let's create a method to load each type of Task into the TaskList
for every branch of the switch case.