Implement Sort Deadline Feature
Added a new feature to sort deadline tasks by their due dates in the TaskList class.

Modified the SortDeadlineCommand to display the sorted list to the user but not saving it to the jarvis.txt yet.

Consider future addition to allow the user to choose to save the filtered order to the storage file or not.