feat: Add flexibility to task search
Previously, the task search functionality required the given
input user keyword string to be an exact match to any substring
of a task name.

This commit allows for better flexibility by:

* firstly, making the search function case-insensitive
* secondly, trimming the keyword input to prevent any leading
  or lagging whitespace to impact sensitivity of the search.

Which are both reasonable mistakes that users may make
when attempting to search for a specific set of tasks.