Add some assertions to some variables
The index of the tasks must be positive integer. The adding event
command must have 2 slashes. The images of user and duke must be not
null.

Add assertions to make sure the program works properly. For example,
it is nonsense to delete the task which has index -1 or 0.

Using assertions is preferable because assertions are used to make sure
that the program works properly given the assumptions about the program
state