Add A-Assertions
Program is unable detect if invalid arguments are passed to the methods.

Program might crash if invalid arguments are
passed to methods unknowingly (e.g. regressions).

Let's add assertion statements in each method
to make sure the parameters passed in are of expected format.

This way regressions and runtime errors are easier identified.