Enhance Friendlier-Syntax extension
The previous version of friendlier-syntax extension was updated for the chatbot to remember the alias setting that the user has set so that users do not have to re-set the alias everytime they start the chatbot.

This enhancement make it more convenient for user to use the chatbot, improving the user experience.