Fix bug on encoding saved tasks
There were observations on different devices that the file duke.txt
containing saved tasks has a newline character at the beginning of the file.

Another observation seen is that there is no newline character after the first task
is encoded, causing the second task to be
written on the same line as the first task

The loadTask() function subsequently reads the line as a corrupted line.
Some reported issues relaunching the app.

This has been corrected and SeeWhyAre app should be launching normally.