(Dummy commit for A-FullCommitMessage on the A-Assertions commit I forgot to do earlier)
Add assertions to various points in the code

The program does not have features to indicate a possible
bug in the code during runtime.

A bug during runtime should be addressed as soon as possible so
that the damage is minimized.

Let's add assertions to document important assumptions that
should hold at various points in the code.

Assertion failures indicates a possible bug and allows us to take
drastic action such as termination of the program.