Add more error handling with more specific exceptions
Previously, parser exceptions were handled at various levels in
the parser functions. Since we now need the exceptions to bubble
to the Atlas::getResponse() method, the exceptions are no longer
caught at earlier levels.

Custom exceptios have also been defined to provide greater
information in debugging, compared to the generic Java exceptions.

The parser now throws an exception if the reminder date is after
the deadline/end of event, while the Event constructor throws an
exception if the start of event is after the end of event.