Change save data to save on every change
Data is written to the file only after the ExitCommand is ran or when
the user closes the interface.

This makes it risky as data can be lost if the ExitCommand is not ran
or the application did not close properly.

Saving the data to the file after every command that causes that data
to change will make it safer as data is saved even if the application
runs into unexpected issues.

Saving data after necessary commands ensure that the data is saved.