Add a UG in the README.md file
To top things off for the iP, the user must, of course,
have a user guide to refer to.

Let's add a UG in the README.md.

Also other files have been updated to clear up bugs,
particularly caused by the Parser class in handling
task editing and trivia handling.