Implement sorting of TaskList by chronological order
TaskList order is sorted by order of task creation.

Such order of the TaskList when displayed is not intuitive.

Let's make the Task class a comparable, and call the 'sort' method of collections when the user displays the tasklist.

The TaskList is now displayed in a user friendly and intuitive format.
Sorting the order of the tasks when the user calls list can reduce the number of time 'sort' is called, hence increasing efficiency.