MainWindow class: include app termination function
User can continue interacting with the chatbot after "bye" was entered.

This deviates from the intended functionality, where the app is supposed
to terminate when "bye" is inputted.

When a "bye" message is received, a Terminate Command is outputted to
the handleUserInput() method in MainWindow, which will trigger
Platform.exit(). This will close the entire app itself.

It is done this way as execute() is supposed to return a String. Hence,
this function is left to MainWindow class to detect for the Terminate
Command instead.