Add showTaskStatics() function
When we have a lot of tasks in the list, it would be convenient to
have a function that shows the user how many each type of task
there are.

Let's have a new user command, "stats", to invoke the
showTaskStatics() function.

It is a convenient way of showing task information to the user.