implement recurring tasks
users have to manually add in weekly repeating tasks.

This process of adding the same task every week is tedious.
The recurring task functionality allow users to add repeating tasks only once, and it will be auto updated when past.

Let's check whether recurring tasks are past when user calls 'list'. If the recurring task is past (i.e. expired), we update the date of the task
to the corresponding day of the week after the current day.

We only display one instance of the recurring task on the task list at any point of time to reduce cluttering.