Improve code quality
Change variable names in DeleteCommand, MarkCommand and UnmarkCommand
to improve readability.

Add JavaDocs to explain code in Bert and Launcher clearly.

Refactor code in the save method of the Storage class. After the
implementation of the GUI, the system output is not needed.
Throws an IOException, which can be caught by the ExitCommand
to produce a CommandResult containing the error.